Moved the code to beta_snippets.py
